Just more of the same

+++ src/niced3d12/fence.rs
@@ -31,11 +31,11 @@ impl SFence {
     }
 
     pub fn wait_for_value(&self, val: u64) {
-        self.wait_for_value_duration(val, <u64>::max_value())
+        self.wait_for_value_duration(val, <u32>::max_value())
             .unwrap();
     }
 
-    pub fn wait_for_value_duration(&self, val: u64, duration: u64) -> Result<(), &'static str> {
+    pub fn wait_for_value_duration(&self, val: u64, duration: u32) -> Result<(), &'static str> {
         if self.raw.getcompletedvalue() < val {
             self.raw.seteventoncompletion(val, &self.fenceevent)?;
             self.fenceevent.waitforsingleobject(duration);
